+++ studypanel/frontend/src/app/pages/active-releases/active-releases.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-active-releases',
+  standalone: true,
+  imports: [],
+  templateUrl: './active-releases.component.html',
+  styleUrl: './active-releases.component.scss',
+})
+export class ActiveReleasesComponent {}