+++ studypanel/frontend/src/app/pages/cloud/cloud.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-cloud',
+  standalone: true,
+  imports: [],
+  templateUrl: './cloud.component.html',
+  styleUrl: './cloud.component.scss',
+})
+export class CloudComponent {}