+++ studypanel/frontend/src/app/pages/rollout/rollout.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-rollout',
+  standalone: true,
+  imports: [],
+  templateUrl: './rollout.component.html',
+  styleUrl: './rollout.component.scss',
+})
+export class RolloutComponent {}