+++ studypanel/frontend/src/app/pages/rollouts/rollouts.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-rollouts',
+  standalone: true,
+  imports: [],
+  templateUrl: './rollouts.component.html',
+  styleUrl: './rollouts.component.scss',
+})
+export class RolloutsComponent {}